fix(runtime): -kagura-anti-tamper could not link anywhere but Apple
AntiTamper.cpp emits kagura_self_check() into main and
kagura_runtime_hash_check() per instrumented function on every target
except Wasm — isWasmTarget is its only guard. Both were defined solely in
ios/jailbreak_detection.c, which runtime/CMakeLists.txt compiles under
if(APPLE). So the pass emitted calls to symbols that do not exist on
Linux, Android or Windows.

The file already documents exactly this shape of gap for Windows
anti-debug a few lines further down. This one went unnoticed because the
only Linux CI job that runs ctest configures with
KAGURA_FORCE_STATIC_PLUGIN=ON, and tests/CMakeLists.txt returns early in
that mode — so the 30 link tests, including link_kagura-tamper, have
never run on Linux.

Neither function is actually OS-specific: the hash check branches on
architecture and on dladdr availability. Both moved to
core/self_check.c in the portable source list, and the aggregate
dispatches to whichever per-platform detectors the target has. Where
there are none (Windows, desktop Linux) it returns 0, documented in place
as "kagura did not detect anything" rather than "known clean" — a
hardcoded 1 would make the pass unusable on the desktop builds
developers test with.

Four Android root detectors — kagura_check_su_binary,
kagura_check_root_packages, kagura_check_test_keys,
kagura_check_rw_system — were also living in that Apple-only file behind
`#if defined(__ANDROID__)`, declared in internal.h and never once
compiled. Confirmed by nm on the pre-change archive: zero matches for all
four, while the two self-check symbols were present. Moved to
android/root_paths.c, deliberately without the __ANDROID__ guard, since
the guard is what hid them; nothing in them is Bionic-specific and they
return 0 on desktop Linux because the paths are absent.

Two bugs surfaced by the move: kagura_check_rw_system returned from
inside its token loop without fclose(), leaking the /proc/mounts
descriptor on exactly the rooted-device path the check exists to detect;
and it used strtok, whose cursor is one process-wide static, in code
AntiTamper can schedule at the entry of any function on any thread.

Also: core/random.c called getrandom() with no API-level guard. Bionic
declares it __INTRODUCED_IN(28), so any build targeting minSdk 21-27
failed to compile. Demonstrated across API levels — 24 and 27 error, 28
and 34 build. The /dev/urandom fallback immediately below already covers
those levels.

Verified by cross-compiling the portable plus Android source lists with
the NDK for aarch64-linux-android24/28/34 and armv7a-linux-androideabi24
with the project's warning flags: clean, no unresolved kagura_ symbols,
all seven target symbols defined. On macOS link_kagura-tamper and the
other 29 link tests pass.
